+++ code/src/cljs/sixsq/nuvla/ui/utils/collapsible_card.cljs
@@ -16,10 +16,10 @@
         more? state-atom]
     (fn [_state-atom]
       (let [label     (@tr (if @more? [:less-details] [:more-details]))
-            icon-name (if @more? "caret down" "caret right")]
+            icon-name (if @more? "fal fa-angle-down" "fal fa-angle-right")]
         [:a {:style    {:cursor "pointer"}
              :on-click #(reset! more? (not @more?))}
-         [ui/Icon {:name icon-name}]
+         [uix/Icon {:name icon-name}]
          label]))))